Remove WML client steps.
Remove everything in `5.0 Deploy the model to Watson Machine Learning`
and below. This is only for use with OpenScale, which is not currently
in use for this workshop.
The actual model save is in section `4.0 Save the model and test data`,
and having secton `5.0` is therefore confusing, so remove it.