Fix broken case.
Linux is case sensitive, so the name of the directory matters, but
git is having this problem:

```bash
 $ git mv machine-learning-in-Jupyter-notebook machine-learning-in-jupyter-notebook
fatal: renaming 'workshop/machine-learning-in-Jupyter-notebook' failed: Invalid argument
```
In order to make the link work, change the case in SUMMARY.md. It would
be better to rename the directory, but git doesn't like that.